Fixed problem with master mode.
The library incorrectly started the slave constructor instead of master.

+++ PCF8574.h
@@ -34,7 +34,7 @@ public:
   explicit PCF8574(const uint8_t deviceAddress = 0x20, TwoWire *wire = &Wire);
 
 #if defined (ESP8266) || defined(ESP32)
-  bool    begin(uint8_t sda, uint8_t scl, uint8_t value = PCF8574_INITIAL_VALUE);
+  bool    begin(int sda, int scl, uint8_t value = PCF8574_INITIAL_VALUE);
 #endif
   bool    begin(uint8_t value = PCF8574_INITIAL_VALUE);
   bool    isConnected();
